refactor: separate PDF inline evidence and share equivalent primitives
