+++ mineru/model/flash/pdf/inline/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) Opendatalab. All rights reserved.
+"""按职责拆分的 Flash PDF 内部实现。"""
+
+__all__: list[str] = []
