+++ mineru/model/flash/pdf/line_layout.py
@@ -13,6 +13,7 @@ from typing import Sequence
 from ....utils.text import is_hyphen_at_line_end
 from ....types import BBox
 
+from .typography import _normalized_font_family
 from .models import (
     _LineItem,
     _LocalAxisLine,
@@ -44,17 +45,6 @@ def _title_fonts_compatible(first: _LineItem, second: _LineItem) -> bool:
     )
     weight_conflicts = _font_weights_conflict(first, second)
     return not (font_conflicts or weight_conflicts)
-
-
-def _normalized_font_family(
-    signature: tuple[str, int] | None,
-) -> str | None:
-    """移除 PDF 字体子集前缀并归一化字体族名称，供几何续行作软兼容判断。"""
-
-    if signature is None:
-        return None
-    name = re.sub(r"^[A-Z]{6}\+", "", signature[0])
-    return re.sub(r"[\s_-]+", "", name).casefold() or None
 
 
 def _font_signatures_share_family(
