+++ mineru/model/flash/pdf/table_recovery/sparse_multiline.py
@@ -9,6 +9,7 @@ from collections import Counter
 from dataclasses import dataclass
 from typing import Any
 
+from .sparse_common import _LocalRule, _local_rules, cluster_members
 from .candidate import GridCellSpec, build_candidate
 from .contracts import (
     NativeTableCandidate,
@@ -26,17 +27,6 @@ from .geometry import (
 )
 
 MIN_MULTILINE_RELIABILITY = 0.98
-
-
-@dataclass(frozen=True, slots=True)
-class _LocalRule:
-    """保存正向表格局部坐标中的线段区间。"""
-
-    orientation: str
-    coordinate: float
-    start: float
-    end: float
-    width: float
 
 
 @dataclass(frozen=True, slots=True)
@@ -65,68 +55,6 @@ class _LogicalRow:
     visual_indices: tuple[int, ...]
     top: float
     bottom: float
-
-
-def _cluster_members(
-    values: list[float],
-    tolerance: float,
-) -> list[tuple[float, tuple[float, ...]]]:
-    """按相邻距离聚类坐标并保留每簇成员。"""
-
-    if not values:
-        return []
-    clusters: list[list[float]] = [[value] for value in sorted(values)[:1]]
-    for value in sorted(values)[1:]:
-        center = float(statistics.median(clusters[-1]))
-        if abs(value - center) <= tolerance:
-            clusters[-1].append(value)
-        else:
-            clusters.append([value])
-    return [(float(statistics.median(cluster)), tuple(cluster)) for cluster in clusters]
-
-
-def _local_rules(
-    table_input: NativeTableInput,
-    width: float,
-    height: float,
-) -> tuple[_LocalRule, ...]:
-    """把页面 drawing 转换为局部横竖线证据。"""
-
-    table_bbox = normalize_bbox(table_input.table_bbox)
-    if table_bbox is None:
-        return ()
-    angle = normalize_angle(table_input.angle)
-    output: list[_LocalRule] = []
-    for rule in table_input.drawing_lines:
-        bbox = normalize_bbox(rule.bbox)
-        if bbox is None:
-            continue
-        local_bbox = page_bbox_to_table_local(bbox, table_bbox, angle)
-        if local_bbox is None:
-            continue
-        local_width = local_bbox[2] - local_bbox[0]
-        local_height = local_bbox[3] - local_bbox[1]
-        if local_width >= max(1.0, 3.0 * max(local_height, 0.1)):
-            output.append(
-                _LocalRule(
-                    orientation="horizontal",
-                    coordinate=(local_bbox[1] + local_bbox[3]) / 2.0,
-                    start=max(0.0, local_bbox[0]),
-                    end=min(width, local_bbox[2]),
-                    width=max(rule.width, local_height),
-                )
-            )
-        elif local_height >= max(1.0, 3.0 * max(local_width, 0.1)):
-            output.append(
-                _LocalRule(
-                    orientation="vertical",
-                    coordinate=(local_bbox[0] + local_bbox[2]) / 2.0,
-                    start=max(0.0, local_bbox[1]),
-                    end=min(height, local_bbox[3]),
-                    width=max(rule.width, local_width),
-                )
-            )
-    return tuple(output)
 
 
 def _local_rectangles(
@@ -268,7 +196,7 @@ def _rectangle_tracks(
 
     edges = [coordinate for rectangle in rectangles for coordinate in (rectangle.bbox[0], rectangle.bbox[2])]
     positions: list[float] = [0.0, width]
-    for coordinate, members in _cluster_members(edges, tolerance):
+    for coordinate, members in cluster_members(edges, tolerance):
         if len(members) < 2:
             continue
         snapped = 0.0 if coordinate <= outer_tolerance else width if width - coordinate <= outer_tolerance else coordinate
@@ -295,7 +223,7 @@ def _filled_band_count(
         if bottom - top < 0.75 * median_height or right - left < 0.08 * width:
             continue
         bands.append((top, bottom))
-    clustered = _cluster_members(
+    clustered = cluster_members(
         [(top + bottom) / 2.0 for top, bottom in bands],
         max(tolerance, 0.50 * median_height),
     )
